Arrange files
